% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-complexheatmap.R
\name{cheat_decorate}
\alias{cheat_decorate}
\title{Decorate ComplexHeatmap}
\usage{
cheat_decorate(vp_name, code)
}
\arguments{
\item{vp_name}{A valid viewport name returned by
\link[ComplexHeatmap:list_components]{list_components}}

\item{code}{Codes to draw elements in the viewport.}
}
\value{
Decorate the viewport.
}
\description{
Decorate ComplexHeatmap
}
\examples{
ggheat(matrix(rnorm(81), nrow = 9), column_km = 2L, name = "cheat_decorate")
ComplexHeatmap::list_components()
cheat_decorate("cheat_decorate_heatmap_body_wrap", {
    grid.text("I'm from cheat_decorate",
        1.5 / 10, 2.5 / 4,
        default.units = "npc"
    )
})
}
