% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganno.R
\name{draw,ggAnnotationFunction-method}
\alias{draw,ggAnnotationFunction-method}
\title{Draw the ggAnnotationFunction Object}
\usage{
\S4method{draw}{ggAnnotationFunction}(object, index, ...)
}
\arguments{
\item{object}{The \code{ggAnnotationFunction-class} object.}

\item{index}{Index of observations.}

\item{...}{Additional arguments passed on to
\link[ComplexHeatmap:draw-HeatmapAnnotation-method]{draw-AnnotationFunction}.}
}
\value{
draw the annotation.
}
\description{
Draw the ggAnnotationFunction Object
}
\examples{
draw(gganno(rnorm(10L), function(p) {
    p + geom_point(aes(y = V1))
}, height = unit(10, "cm"), width = unit(0.7, "npc")))
}
