% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-methods.R, R/eanno.R
\name{draw-method}
\alias{draw-method}
\alias{draw}
\alias{draw,gtable-method}
\alias{draw,ggplot-method}
\alias{draw,ExtendedAnnotation-method}
\title{Draw methods for additional objects}
\usage{
draw(object, ...)

\S4method{draw}{gtable}(object)

\S4method{draw}{ggplot}(object)

\S4method{draw}{ExtendedAnnotation}(object, index, k = 1L, n = 1L, ...)
}
\arguments{
\item{object}{See method signature}

\item{...}{Additional arguments passed on to}

\item{index}{A vector of indices.}

\item{k}{The current slice index for the annotation if it is split.}

\item{n}{Total number of slices.}
}
\value{
draw the annotation.
}
\description{
Only object with \link{make_legends} methods can be put in
\code{legends_margin}. Only object with \link[=draw-method]{draw} methods can be put in
\code{legends_panel}.
}
\examples{
draw(eheat(matrix(rnorm(81), nrow = 9)))
}
