% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eheat-package.R
\docType{package}
\name{eheat-package}
\alias{eheat-package}
\title{eheat: Extended ComplexHeatmap}
\description{
This package serves as a bridge between the ggplot2 and ComplexHeatmap packages. Essentially, all ggplot2 geometries and operations can be utilized in ComplexHeatmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/eheat}
  \item Report bugs at \url{https://github.com/Yunuuuu/eheat/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
