% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eheat_scale.R
\name{eheat_map}
\alias{eheat_map}
\title{Wrap ggplot2 scale object into ComplexHeatmap}
\usage{
eheat_map(scale = NULL, aesthetic = NULL, matrix = NULL)
}
\arguments{
\item{scale}{A \link[ggplot2:discrete_scale]{scale} object.}

\item{matrix}{Matrix used for this aesthetic mapping}
}
\description{
Wrap ggplot2 scale object into ComplexHeatmap
}
