% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganno.R, R/ggheat.R, R/prepare.R
\name{eheat_prepare.ggAnno}
\alias{eheat_prepare.ggAnno}
\alias{eheat_prepare.ggHeatmap}
\alias{eheat_prepare}
\alias{eheat_prepare.ExtendedHeatmap}
\alias{eheat_prepare.ExtendedAnnotation}
\title{Prepare ExtendedHeatmap}
\usage{
\method{eheat_prepare}{ggAnno}(object, ..., viewport, heatmap, name)

\method{eheat_prepare}{ggHeatmap}(object, ...)

eheat_prepare(object, ...)

\method{eheat_prepare}{ExtendedHeatmap}(object, ...)

\method{eheat_prepare}{ExtendedAnnotation}(object, ..., viewport, heatmap, name)
}
\arguments{
\item{object}{A \link[=eheat]{ExtendedHeatmap} or \link[=eanno]{ExtendedAnnotation}
object.}

\item{...}{Not used currently.}

\item{viewport}{A viewport for this annotation.}

\item{heatmap}{Heatmap object after clustering.}

\item{name}{A string, the name of the annotation.}
}
\value{
An modified \code{object} with the same class.
}
\description{
Prepare ExtendedHeatmap
}
\examples{
eheat_prepare(eheat(matrix(rnorm(81), nrow = 9)))
}
