% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{get_guides}
\alias{get_guides}
\alias{get_guides.default}
\alias{get_guides.ggplot}
\alias{get_guides.gtable}
\title{Get guide legends as a list of \code{gtable} object}
\usage{
get_guides(x, ...)

\method{get_guides}{default}(x, ...)

\method{get_guides}{ggplot}(x, ...)

\method{get_guides}{gtable}(x, margins = NULL, ...)
}
\arguments{
\item{x}{See method signatures}

\item{...}{Not used currently.}

\item{margins}{Guides of which margin to draw besides the plot in the \code{vp}
viewport. Allowed values are \code{"t"}, \code{"l"}, \code{"b"}, \code{"r"}, and \code{"i"}. \code{"i"} means
the inside legends (\code{ggplot2::theme(legend.position = "inside")}). Default:
\code{NULL}, which indicates \code{c("t", "l", "b", "r")}.}
}
\value{
A list of \link[gtable:gtable]{gtable} objects.
}
\description{
Get guide legends as a list of \code{gtable} object
}
\examples{
gg <- ggplot(mtcars) +
    geom_bar(
        aes(mpg, disp, fill = factor(mpg)),
        stat = "identity"
    )
get_guides(gg)
}
