% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganno.R
\docType{class}
\name{gganno}
\alias{gganno}
\alias{ggAnnotationFunction-class}
\title{Build AnnotationFunction Class with ggplot2}
\usage{
gganno(
  matrix,
  ggfn,
  ...,
  which = NULL,
  width = NULL,
  height = NULL,
  debug = FALSE
)
}
\arguments{
\item{matrix}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{ggfn}{A function or formula, accept a initial \link[ggplot2:ggplot]{ggplot}
data as the input and must return a \link[ggplot2:ggplot]{ggplot} object.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function with up to
two arguments: \code{.x} (single argument) or \code{.x} and \code{.y} (two arguments). The
\code{.} placeholder can be used instead of \code{.x}.  This allows you to create
very compact anonymous functions (lambdas) with up to two inputs.}

\item{...}{Other arguments passed to \code{ggfn}.}

\item{which}{Whether it is drawn as a column annotation or a row annotation?}

\item{width}{The width of the plotting region (the viewport) that the annotation is drawn. If it is a row annotation, the width must be an absolute unit. Since the \code{AnnotationFunction} object is always contained by the \code{\link[ComplexHeatmap]{SingleAnnotation-class}}object, you can only set the width of row annotations or height of column annotations, while e.g. the height of the row annotation is always \code{unit(1, "npc")} which means it always fully filled in the parent \code{SingleAnnotation} and only in \code{\link[ComplexHeatmap]{SingleAnnotation}} or even \code{\link[ComplexHeatmap]{HeatmapAnnotation}} can adjust the height of the row annotations.}

\item{height}{The height of the plotting region (the viewport) that the annotation is drawn. If it is a column annotation, the width must be an absolute unit.}

\item{debug}{In the event that it is set to TRUE, the
\link[ggplot2:ggplot]{ggplot} object gracefully yielded by ggfn, during the
execution of \link[ComplexHeatmap:draw-dispatch]{draw}, shall be returned directly.
Alternatively, one may opt to impart a function (whereas a formula would also
be deemed acceptable) to be carried out conjointly with the
\link[ggplot2:ggplot]{ggplot} object, as exemplified by \link{print} or \link{browser}.}
}
\value{
A \code{ggAnnotationFunction} object.
}
\description{
Build AnnotationFunction Class with ggplot2
}
