% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfit-.R
\name{ggfit}
\alias{ggfit}
\title{Fit ggplot2 panel or plot in a viewport}
\usage{
ggfit(
  gg,
  align = "panel",
  vp = NULL,
  sides = c("b", "t", "l", "r"),
  elements = c("axis", "lab", "guide"),
  background = FALSE,
  gt = NULL
)
}
\arguments{
\item{gg}{A \link[ggplot2:ggplot]{ggplot2} object.}

\item{align}{A string indicates how to align the \code{viewport}. Must be "panel"
or "plot".}

\item{vp}{A \link[grid:viewport]{viewport} object.}

\item{sides}{Which side to draw besides the panel elements. If \code{NULL}, will
draw panel only.}

\item{elements}{Ggplot elements to draw, can be a list of a character to
specify elements for each side separately. Valid elements are "axis", "lab",
"guide". Other elements will be ignored.}

\item{background}{A boolean value indicates whether to draw the background.}

\item{gt}{A \link[ggplot2:ggplotGrob]{gtable} object.}
}
\value{
Draw ggplot object.
\itemize{
\item align = \code{"panel"}: Draw ggplot object by fitting exactly the panel to \code{vp}.
\item align = \code{"plot"}: Draw ggplot object in \code{vp}.
}
}
\description{
Fit ggplot2 panel or plot in a viewport
}
\examples{
p <- ggplot(data.frame(x = 0:10, y = 0:10), aes(x, y)) +
    geom_point()
outerBox <- viewport(width = unit(125, "mm"), height = unit(150, "mm"))
innerBox <- viewport(
    x = unit(0.5, "npc"), y = unit(0.6, "npc"),
    width = unit(60, "mm"), height = unit(70, "mm"), angle = -30
)

# ggfit_panel ------------
grid.newpage()
pushViewport(outerBox)
grid.rect(gp = gpar(col = "red", fill = NA))

pushViewport(innerBox)
grid.rect(gp = gpar(col = "red", fill = NA, lwd = 2))
ggfit(p, "panel")

# ggfit_plot -------------
grid.newpage()
pushViewport(outerBox)
grid.rect(gp = gpar(col = "red", fill = NA))

pushViewport(innerBox)
grid.rect(gp = gpar(col = "red", fill = NA, lwd = 2))
ggfit(p, "plot")
}
\seealso{
\itemize{
\item \link{ggfit_panel}
\item \link{ggfit_plot}
}
}
