% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\docType{class}
\name{ggheat}
\alias{ggheat}
\alias{ggHeatmap-class}
\title{Extended Heatmap}
\usage{
ggheat(matrix, ggfn = NULL, ..., ggparams = list())
}
\arguments{
\item{matrix}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{ggfn}{A function or formula, accept a initial \link[ggplot2:ggplot]{ggplot}
data as the input and must return a \link[ggplot2:ggplot]{ggplot} object. For the
passed data details, please see section \code{ggfn}.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function with up to
two arguments: \code{.x} (single argument) or \code{.x} and \code{.y} (two arguments). The
\code{.} placeholder can be used instead of \code{.x}.  This allows you to create
very compact anonymous functions (lambdas) with up to two inputs.}

\item{...}{Additional arguments passed on to \link{eheat}.}

\item{ggparams}{Other arguments passed to \code{ggfn}.}
}
\value{
A \code{ggHeatmap} Object.
}
\description{
Plot heatmaps layer by layer
}
\note{
Maintaining the internal limits along the heatmap to align well with
\code{ComplexHeatmap} is important.
}
\section{ggfn}{


\code{ggfn} accept a ggplot2 object with a default data and mapping created by
\code{ggplot(data, aes(.data$x, .data$y))}.  the data contains following columns:
\itemize{
\item \code{.slice}: slice number, combine \code{.slice_row} and \code{.slice_column}.
\item \code{.slice_row}: the slice row number
\item \code{.slice_column}: the slice column number
\item \code{.row_names} and \code{.column_names}: the row and column names of the original
matrix (only applicable when names exist).
\item \code{.row_index} and \code{.column_index}: the row and column index of the original
matrix.
\item \code{x} and \code{y}: the \code{x} and \code{y} coordinates
\item \code{value}: the actual matrix value for the heatmap matrix.
}
}

\examples{
ggheat(matrix(rnorm(81), nrow = 9))
}
