% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eanno.R, R/eheat.R, R/internal-methods.R
\name{make_layout,ExtendedAnnotation-method}
\alias{make_layout,ExtendedAnnotation-method}
\alias{prepare,ExtendedHeatmap-method}
\alias{make_layout,ExtendedHeatmap-method}
\alias{make_layout,HeatmapAnnotation-method}
\alias{internal-method}
\title{Internal methods to extend ComplexHeatmap}
\usage{
\S4method{make_layout}{ExtendedAnnotation}(object, ..., viewport = NULL, heatmap = NULL, name = NULL)

\S4method{prepare}{ExtendedHeatmap}(object, process_rows = TRUE, process_columns = TRUE)

\S4method{make_layout}{ExtendedHeatmap}(object)

\S4method{make_layout}{HeatmapAnnotation}(object, heatmap)
}
\arguments{
\item{object}{See method signatures.}

\item{viewport}{A viewport for this annotation.}

\item{heatmap}{Heatmap object after clustering.}

\item{name}{A string, the name of the annotation.}

\item{process_rows}{Whether to process rows of the heatmap.}

\item{process_columns}{Whether to process columns of the heatmap.}
}
\value{
A modified object with the same class.
}
\description{
Internal methods to extend ComplexHeatmap
}
\examples{
prepare(eheat(matrix(rnorm(81), nrow = 9)))
}
\keyword{internal}
