% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_from_gg}
\alias{legend_from_gg}
\title{Legends from ggplot2}
\usage{
legend_from_gg(gg)
}
\arguments{
\item{gg}{A \link[ggplot2:ggplot]{ggplot2} object.}
}
\value{
A list of \link[ComplexHeatmap:Legends]{Legends} object.
}
\description{
Legends from ggplot2
}
\examples{
gg <- ggplot(mtcars) +
    geom_bar(
        aes(mpg, disp, fill = factor(mpg)),
        stat = "identity"
    )
draw(legend_from_gg(gg)[[1L]])
}
\seealso{
\link[ComplexHeatmap:Legend]{ComplexHeatmap::Legend}
}
