% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{make_legends}
\alias{make_legends}
\title{Legends from ggplot2}
\usage{
make_legends(x, margins = NULL)
}
\arguments{
\item{x}{A \link[ggplot2:ggplot]{ggplot2} object or a \link[gtable:gtable]{gtable}
object.}

\item{margins}{Guides of which margin to draw besides the plot in the \code{vp}
viewport. Allowed values are \code{"t"}, \code{"l"}, \code{"b"}, \code{"r"}, and \code{"i"}. "i" means the
inside legends (\code{ggplot2::theme(legend.position = "inside")}). Default:
\code{NULL}, which indicates \code{c("t", "l", "b", "r")}.}
}
\value{
A list of \link[ComplexHeatmap:Legends]{Legends} object.
}
\description{
Legends from ggplot2
}
\examples{
gg <- ggplot(mtcars) +
    geom_bar(
        aes(mpg, disp, fill = factor(mpg)),
        stat = "identity"
    )
draw(make_legends(gg)[[1L]])
}
\seealso{
\link[ComplexHeatmap:Legend]{ComplexHeatmap::Legend}
}
