% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare,ggHeatmap-method}
\alias{prepare,ggHeatmap-method}
\title{Prepare the Heatmap}
\usage{
\S4method{prepare}{ggHeatmap}(object, process_rows = TRUE, process_columns = TRUE)
}
\arguments{
\item{object}{A \code{\link[ComplexHeatmap]{Heatmap-class}} object.}

\item{process_rows}{Whether to process rows of the heatmap.}

\item{process_columns}{Whether to process columns of the heatmap.}
}
\value{
The \code{\link[ComplexHeatmap]{Heatmap-class}} object.
}
\description{
Prepare the Heatmap
}
\details{
The preparation of the heatmap includes following steps:

\itemize{
  \item making clustering on rows (by calling \code{\link[ComplexHeatmap]{make_row_cluster,Heatmap-method}})
  \item making clustering on columns (by calling \code{\link[ComplexHeatmap]{make_column_cluster,Heatmap-method}})
  \item making the layout of the heatmap (by calling \code{\link[ComplexHeatmap]{make_layout,Heatmap-method}})
}

This function is only for internal use.
}
\examples{
# There is no example
NULL
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
