% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gpar}
\alias{unit}
\alias{alpha}
\alias{.pt}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:graphical-units]{.pt}}, \code{\link[ggplot2:reexports]{alpha}}}

  \item{grid}{\code{\link[grid]{gpar}}, \code{\link[grid]{unit}}}
}}

