% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-complexheatmap.R
\name{with_ht_opts}
\alias{with_ht_opts}
\title{Temporarily change ComplexHeatmap options.}
\usage{
with_ht_opts(opts, code)
}
\arguments{
\item{opts}{New \link[ComplexHeatmap:ht_opt]{ht_opt}, must be a named list.}

\item{code}{Running \link{ggheat} or \link[ComplexHeatmap:Heatmap]{Heatmap} function.}
}
\value{
The results of the evaluation of the \code{code} argument.
}
\description{
Temporarily change ComplexHeatmap options.
}
\note{
ComplexHeatmap heatmap options will always be reset with
\code{ComplexHeatmap::ht_opt(RESET = TRUE)} when function exit.
}
\examples{
with_ht_opts(
    list(legend_title_gp = gpar(col = "red")),
    ggheat(matrix(rnorm(81), nrow = 9))
)
with_ht_opts(
    list(verbose = TRUE),
    ggheat(matrix(rnorm(81), nrow = 9))
)
}
\seealso{
\link[ComplexHeatmap:ht_opt]{ht_opt}
}
