% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-complexheatmap.R
\name{with_ht_verbose}
\alias{with_ht_verbose}
\title{Make ComplexHeatmap verbose}
\usage{
with_ht_verbose(code)
}
\arguments{
\item{code}{Running \link{ggheat} or \link[ComplexHeatmap:Heatmap]{Heatmap} function.}
}
\value{
The results of the evaluation of the \code{code} argument.
}
\description{
Make ComplexHeatmap verbose
}
\examples{
with_ht_verbose(ggheat(matrix(rnorm(81), nrow = 9)))
}
\seealso{
\link{with_ht_opts}
}
