% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno-gg.R
\name{anno_gg}
\alias{anno_gg}
\alias{anno_gg2}
\title{Create a heatmap annotation with ggplot object.}
\usage{
anno_gg(
  gg,
  align_with = "full",
  clip = NULL,
  gt = NULL,
  width = NULL,
  height = NULL,
  show_name = FALSE,
  which = NULL
)

anno_gg2(
  gg,
  align_with = "full",
  margins = c("b", "t", "l", "r"),
  elements = c("axis", "lab", "guide", "subtitle", "title", "caption"),
  gt = NULL,
  width = NULL,
  height = NULL,
  show_name = FALSE,
  which = NULL
)
}
\arguments{
\item{gg}{A \link[ggplot2:ggplot]{ggplot2} object.}

\item{align_with}{Which area of the object should be fit to the viewport?
Must be a string of "panel", "plot", "full".}

\item{clip}{One of \code{"on"}, \code{"inherit"}, or \code{"off"}, indicating whether to
clip to the extent out of the viewport, inherit the clipping region from
\code{vp}, or turn clipping off altogether. A logical value of \code{TRUE} corresponds
to \code{"on"} and \code{FALSE} corresponds to \code{"off"}. Default: \code{"inherit"}.}

\item{gt}{A \link[ggplot2:ggplotGrob]{gtable} object.}

\item{width}{The width of the plotting region (the viewport) that the annotation is drawn. If it is a row annotation, the width must be an absolute unit. Since the \code{AnnotationFunction} object is always contained by the \code{\link[ComplexHeatmap]{SingleAnnotation-class}}object, you can only set the width of row annotations or height of column annotations, while e.g. the height of the row annotation is always \code{unit(1, "npc")} which means it always fully filled in the parent \code{SingleAnnotation} and only in \code{\link[ComplexHeatmap]{SingleAnnotation}} or even \code{\link[ComplexHeatmap]{HeatmapAnnotation}} can adjust the height of the row annotations.}

\item{height}{The height of the plotting region (the viewport) that the annotation is drawn. If it is a column annotation, the width must be an absolute unit.}

\item{show_name}{It is used to turn off the drawing of annotation names in \code{\link[ComplexHeatmap]{HeatmapAnnotation}}. Annotations always have names associated and normally they will be drawn beside the annotation graphics to tell what the annotation is about. e.g. the annotation names put beside the points annotation graphics. However, for some of the annotations, the names are not necessarily to be drawn, such as text annotations drawn by \code{\link[ComplexHeatmap]{anno_text}} or an empty annotation drawn by \code{\link[ComplexHeatmap]{anno_empty}}. In this case, when \code{show_names} is set to \code{FALSE}, there will be no annotation names drawn for the annotation.}

\item{which}{Whether it is drawn as a column annotation or a row annotation?}

\item{margins}{Which margin to draw besides the plot in the \code{vp} viewport.
This allows for a more precise adjustment of the \code{clip} feature. Allowed
values are \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"}, When set to \code{NULL}, it means clip =
\code{"off"}.}

\item{elements}{Ggplot elements to draw, can be a list of a character to
specify elements for each side separately. Valid elements are
\code{"axis"}, \code{"lab"}, \code{"guide"}, \code{"subtitle"}, \code{"title"}, and \code{"caption"}. Invalid elements will be just omitted. IF
\code{NULL}, all valid ggplot elements will be used.}
}
\value{
A \code{\link[ComplexHeatmap]{AnnotationFunction-class}} object which can be used in \code{\link[ComplexHeatmap]{HeatmapAnnotation}}.
}
\description{
Create a heatmap annotation with ggplot object.
}
\examples{
g <- ggplot(mpg, aes(displ, hwy, colour = class)) +
    geom_point()
m <- matrix(rnorm(100), 10)

# anno_gg-panel: clip = "off" -------
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg(g, "panel",
            clip = "off",
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)

# anno_gg-panel: clip = "on" --------
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg(g, "panel",
            clip = "on",
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)

# anno_gg-plot ---------------------
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg(g, "plot",
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)

# anno_gg-full --------------------
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg(g, "full",
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)
# anno_gg2-panel: margins = NULL -------
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg2(g, "panel",
            margins = NULL,
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)

# anno_gg2-panel: margins = "l" --------
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg2(g, "panel",
            margins = "l",
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)

# anno_gg2-panel: margins = "r" --------
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg2(g, "panel",
            margins = "r",
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)

# anno_gg2-plot ---------------------
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg2(g, "plot",
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)

# anno_gg2-full --------------------
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg2(g, "full",
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)
}
\seealso{
\link{ggfit}
}
