% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno-gg.R
\name{anno_gg}
\alias{anno_gg}
\title{Create a heatmap annotation with ggplot object.}
\usage{
anno_gg(
  gg,
  align = "panel",
  sides = c("b", "t", "l", "r"),
  elements = c("axis", "lab", "guide"),
  background = FALSE,
  gt = NULL,
  width = NULL,
  height = NULL,
  show_name = FALSE,
  which = NULL,
  name = NULL
)
}
\arguments{
\item{gg}{A \link[ggplot2:ggplot]{ggplot2} object.}

\item{align}{A string indicates how to align the \code{viewport}. Must be "panel"
or "plot".}

\item{sides}{Which side to draw besides the panel elements. If \code{NULL}, will
draw panel only.}

\item{elements}{Ggplot elements to draw, can be a list of a character to
specify elements for each side separately. Valid elements are "axis", "lab",
"guide". Other elements will be ignored.}

\item{background}{A boolean value indicates whether to draw the background.}

\item{gt}{A \link[ggplot2:ggplotGrob]{gtable} object.}

\item{width}{The width of the plotting region (the viewport) that the annotation is drawn. If it is a row annotation, the width must be an absolute unit. Since the \code{AnnotationFunction} object is always contained by the \code{\link[ComplexHeatmap]{SingleAnnotation-class}}object, you can only set the width of row annotations or height of column annotations, while e.g. the height of the row annotation is always \code{unit(1, "npc")} which means it always fully filled in the parent \code{SingleAnnotation} and only in \code{\link[ComplexHeatmap]{SingleAnnotation}} or even \code{\link[ComplexHeatmap]{HeatmapAnnotation}} can adjust the height of the row annotations.}

\item{height}{The height of the plotting region (the viewport) that the annotation is drawn. If it is a column annotation, the width must be an absolute unit.}

\item{show_name}{It is used to turn off the drawing of annotation names in \code{\link[ComplexHeatmap]{HeatmapAnnotation}}. Annotations always have names associated and normally they will be drawn beside the annotation graphics to tell what the annotation is about. e.g. the annotation names put beside the points annotation graphics. However, for some of the annotations, the names are not necessarily to be drawn, such as text annotations drawn by \code{\link[ComplexHeatmap]{anno_text}} or an empty annotation drawn by \code{\link[ComplexHeatmap]{anno_empty}}. In this case, when \code{show_names} is set to \code{FALSE}, there will be no annotation names drawn for the annotation.}

\item{which}{Whether it is drawn as a column annotation or a row annotation?}

\item{name}{Name of the annotation, only used for message.}
}
\value{
A \code{\link[ComplexHeatmap]{AnnotationFunction-class}} object which can be used in \code{\link[ComplexHeatmap]{HeatmapAnnotation}}.
}
\description{
Create a heatmap annotation with ggplot object.
}
\examples{
g <- ggplot(mpg, aes(displ, hwy, colour = class)) +
    geom_point()
m <- matrix(rnorm(100), 10)
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg(g, "panel",
            background = TRUE,
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)
ggheat(m,
    top_annotation = HeatmapAnnotation(
        ggplot = anno_gg(g, "plot",
            background = TRUE,
            height = unit(6, "cm"),
            show_name = FALSE
        )
    )
)
}
\seealso{
\link{ggfit}
}
