% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganno.R
\name{draw,ggAnnotationFunction-method}
\alias{draw,ggAnnotationFunction-method}
\title{Draw the ggAnnotationFunction Object}
\usage{
\S4method{draw}{ggAnnotationFunction}(object, index, ...)
}
\arguments{
\item{object}{The \code{ggAnnotationFunction-class} object.}

\item{index}{Index of observations.}

\item{...}{
  Arguments passed on to \code{\link[grid:viewport]{grid::viewport}}
  \describe{
    \item{\code{x}}{A numeric vector or unit object specifying x-location.}
    \item{\code{y}}{A numeric vector or unit object specifying y-location.}
    \item{\code{default.units}}{A string indicating the default units to use
    if \code{x}, \code{y}, \code{width}, or \code{height}
    are only given as numeric vectors.}
    \item{\code{just}}{A string or numeric
    vector specifying the justification of the viewport
    relative to its (x, y) location.  If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left alignment
    and 1 means right alignment.
  }
    \item{\code{gp}}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}
    \item{\code{clip}}{One of \code{"on"}, \code{"inherit"}, or
    \code{"off"}, indicating whether to
    clip to the extent of this viewport, inherit the clipping region
    from the parent viewport, or turn clipping off altogether.
    For back-compatibility, a logical value of \code{TRUE} corresponds
    to \code{"on"} and \code{FALSE} corresponds to \code{"inherit"}.

    May also be a grob (or a gTree) that describes a clipping path
    or the result of a call to \code{\link[grid]{as.path}}.}
    \item{\code{mask}}{One of \code{"none"} (or \code{FALSE})
    or \code{"inherit"} (or \code{TRUE}) 
    or a grob (or a gTree) or the result of call to
    \code{\link[grid]{as.mask}}.  This specifies that the viewport
    should have no mask, or it should inherit the mask of its parent,
    or it should have its own mask, as described by the grob.}
    \item{\code{xscale}}{A numeric vector of length two indicating the minimum and
    maximum on the x-scale.  The limits may not be identical. }
    \item{\code{yscale}}{A numeric vector of length two indicating the minimum
    and maximum on the y-scale. The limits may not be identical. }
    \item{\code{angle}}{A numeric value indicating the angle of rotation of the
    viewport.  Positive values indicate the amount of rotation, in
    degrees, anticlockwise from the positive x-axis.}
    \item{\code{layout}}{A Grid layout object which splits the viewport into
    subregions.}
    \item{\code{layout.pos.row}}{A numeric vector giving the
    rows occupied by this viewport in its
    parent's layout.}
    \item{\code{layout.pos.col}}{A numeric vector giving the
    columns occupied by this viewport in its
    parent's layout.}
    \item{\code{name}}{A character value to uniquely identify the viewport
    once it has been pushed onto the viewport tree.  }
  }}
}
\description{
Draw the ggAnnotationFunction Object
}
