% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganno.R
\name{draw,ggAnnotationFunction-method}
\alias{draw,ggAnnotationFunction-method}
\title{Draw the ggAnnotationFunction Object}
\usage{
\S4method{draw}{ggAnnotationFunction}(object, index, k = 1L, n = 1L, ...)
}
\arguments{
\item{object}{The \link[=gganno]{ggAnnotationFunction} object.}

\item{index}{A vector of indices.}

\item{k}{The current slice index for the annotation if it is split.}

\item{n}{Total number of slices.}

\item{...}{Additional arguments passed on to}
}
\value{
draw the annotation.
}
\description{
Draw the ggAnnotationFunction Object
}
\examples{
draw(gganno(rnorm(10L), function(p) {
    p + geom_point(aes(y = V1))
}, height = unit(10, "cm"), width = unit(0.7, "npc")))
}
