% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eheat-package.R
\docType{package}
\name{eheat-package}
\alias{eheat-package}
\alias{_PACKAGE}
\title{eheat: Extented ComplexHeatmap}
\description{
Extented ComplexHeatmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/eheat}
  \item Report bugs at \url{https://github.com/Yunuuuu/eheat/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@gmail.com}

}
\keyword{internal}
