% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R, R/gganno.R, R/prepare.R
\name{eheat_prepare.ggHeatmap}
\alias{eheat_prepare.ggHeatmap}
\alias{eheat_prepare.ggAnno}
\alias{eheat_prepare}
\alias{eheat_prepare.ExtendedHeatmap}
\alias{eheat_prepare.ExtendedAnnotation}
\title{Prepare ExtendedHeatmap}
\usage{
\method{eheat_prepare}{ggHeatmap}(object, ...)

\method{eheat_prepare}{ggAnno}(object, order_list, name = NULL, ...)

eheat_prepare(object, ...)

\method{eheat_prepare}{ExtendedHeatmap}(object, ...)

\method{eheat_prepare}{ExtendedAnnotation}(object, order_list, name = NULL, ...)
}
\arguments{
\item{object}{A \link[=eheat]{ExtendedHeatmap} or \link[=eanno]{ExtendedAnnotation}
object.}

\item{...}{Additional arguments passed to specific methods}

\item{order_list}{Heatmap order list (column/row) after clustering.}

\item{name}{A string, the name of the annotation.}
}
\value{
An modified \link[=eheat]{ExtendedHeatmap} or \link[=eanno]{ExtendedAnnotation}
object.
}
\description{
Prepare ExtendedHeatmap
}
\examples{
eheat_prepare(eheat(matrix(rnorm(81), nrow = 9)))
}
