% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\docType{class}
\name{ggheat}
\alias{ggheat}
\alias{ggHeatmap-class}
\title{Extended Heatmap}
\usage{
ggheat(matrix, ggfn = NULL, ..., ggparams = list(), debug = FALSE)
}
\arguments{
\item{matrix}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{ggfn}{A function or formula, accept a initial \link[ggplot2:ggplot]{ggplot}
data as the input and must return a \link[ggplot2:ggplot]{ggplot} object. For the
passed data details, please see section \code{ggfn}.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function with up to
two arguments: \code{.x} (single argument) or \code{.x} and \code{.y} (two arguments). The
\code{.} placeholder can be used instead of \code{.x}.  This allows you to create
very compact anonymous functions (lambdas) with up to two inputs.}

\item{...}{Other arguments passsed to \link[ComplexHeatmap:Heatmap]{Heatmap}.
\itemize{
\item \code{name}: Name of the heatmap. By default the heatmap name is used as the
title of the heatmap legend.
\item \code{border}: Whether draw border. The value can be logical or a string of
color.
\item \code{border_gp}: Graphic parameters for the borders. If you want to set
different parameters for different heatmap slices, please consider to use
\code{decorate_heatmap_body}.
\item \code{cell_fun}: Self-defined function to add graphics on each cell. Seven
parameters will be passed into this function: \code{j}, \code{i}, \code{x}, \code{y},
\code{width}, \code{height}, \code{fill} which are column index, row index in
\code{matrix}, coordinate of the cell, the width and height of the cell and the
filled color. \code{x}, \code{y}, \code{width} and \code{height} are all \code{grid::unit}
objects.  -layer_fun Similar as \code{cell_fun}, but is vectorized. Check
\url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#customize-the-heatmap-body}.
\item \code{jitter}: Random shifts added to the matrix. The value can be logical or a
single numeric value. It it is \code{TRUE}, random values from uniform
distribution between 0 and 1e-10 are generated. If it is a numeric
value, the range for the uniform distribution is (0, \code{jitter}). It is
mainly to solve the problem of "Error: node stack overflow" when there
are too many identical rows/columns for plotting the dendrograms.  ADD:
From version 2.5.6, the error of node stack overflow has been fixed, now
this argument is ignored.
\item \code{row_title}: Title on the row.
\item \code{row_title_side}: Will the title be put on the left or right of the
heatmap?
\item \code{row_title_gp}: Graphic parameters for row title.
\item \code{row_title_rot}: Rotation of row title.
\item \code{column_title} Title on the column.
\item \code{column_title_side} Will the title be put on the top or bottom of the
heatmap?
\item \code{column_title_gp} Graphic parameters for column title.
\item \code{column_title_rot} Rotation of column titles.
\item \code{cluster_rows} If the value is a logical, it controls whether to make
cluster on rows. The value can also be a
\link[stats:hclust]{hclust} or a \code{stats::dendrogram} which already
contains clustering.  Check
\url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#clustering}.
\item \code{cluster_row_slices} If rows are split into slices, whether perform
clustering on the slice means?
\item \code{clustering_distance_rows}: It can be a pre-defined character which is in
("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski",
"pearson", "spearman", "kendall"). It can also be a function.
\itemize{
\item If the function has one argument, the input argument should be a matrix
and the returned value should be a \code{stats::dist} object.
\item If the function has two arguments, the input arguments are two vectors
and the function calculates distance between these two vectors.
}
\item \code{clustering_method_rows} Method to perform hierarchical clustering, pass to
\link[stats:hclust]{hclust}.
\item \code{row_dend_side} Should the row dendrogram be put on the left or right of
the heatmap?
\item \code{row_dend_width} Width of the row dendrogram, should be a \code{grid::unit}
object.
\item \code{show_row_dend} Whether show row dendrogram?
\item \code{row_dend_gp} Graphic parameters for the dendrogram segments. If users
already provide a \code{stats::dendrogram} object with edges rendered, this
argument will be ignored.  -row_dend_reorder Apply reordering on row
dendrograms. The value can be a logical value or a vector which contains
weight which is used to reorder rows. The reordering is applied by
\code{stats::reorder.dendrogram}.
\item \code{cluster_columns} Whether make cluster on columns? Same settings as
\code{cluster_rows}.
\item \code{cluster_column_slices} If columns are split into slices, whether perform
clustering on the slice means?
\item \code{clustering_distance_columns} Same setting as \code{clustering_distance_rows}.
\item \code{clustering_method_columns} Method to perform hierarchical clustering, pass
to \link[stats:hclust]{hclust}.
\item \code{column_dend_side} Should the column dendrogram be put on the top or bottom
of the heatmap?
\item \code{column_dend_height} height of the column cluster, should be a \code{grid::unit}
object.
\item \code{show_column_dend} Whether show column dendrogram?
\item \code{column_dend_gp} Graphic parameters for dendrogram segments. Same settings
as \code{row_dend_gp}.
\item \code{column_dend_reorder} Apply reordering on column dendrograms. Same settings
as \code{row_dend_reorder}.
\item \code{row_order} Order of rows. Manually setting row order turns off clustering.
\item \code{column_order} Order of column.
\item \code{row_labels} Optional row labels which are put as row names in the heatmap.
\item \code{row_names_side} Should the row names be put on the left or right of the
heatmap?
\item \code{show_row_names} Whether show row names.
\item \code{row_names_max_width} Maximum width of row names viewport.
\item \code{row_names_gp} Graphic parameters for row names.
\item \code{row_names_rot} Rotation of row names.
\item \code{row_names_centered} Should row names put centered?
\item \code{column_labels} Optional column labels which are put as column names in the
heatmap.
\item \code{column_names_side} Should the column names be put on the top or bottom of
the heatmap?
\item \code{column_names_max_height} Maximum height of column names viewport.
\item \code{show_column_names} Whether show column names.
\item \code{column_names_gp} Graphic parameters for drawing text.
\item \code{column_names_rot} Rotation of column names.
\item \code{column_names_centered} Should column names put centered?
\item \code{top_annotation} A \link[ComplexHeatmap:HeatmapAnnotation]{HeatmapAnnotation}
object.
\item \code{bottom_annotation} A
\link[ComplexHeatmap:HeatmapAnnotation]{HeatmapAnnotation} object.
\item \code{left_annotation} It should be specified by \code{rowAnnotation}.
\item \code{right_annotation} it should be specified by \code{rowAnnotation}.
\item \code{km} Apply k-means clustering on rows. If the value is larger than 1, the
heatmap will be split by rows according to the k-means clustering.  For
each row slice, hierarchical clustering is still applied with parameters
above.
\item \code{split} A vector or a data frame by which the rows are split. But if
\code{cluster_rows} is a clustering object, \code{split} can be a single
number indicating to split the dendrogram by \code{stats::cutree}.
\item \code{row_km} Same as \code{km}.
\item \code{row_km_repeats} Number of k-means runs to get a consensus k-means
clustering. Note if \code{row_km_repeats} is set to more than one, the final
number of groups might be smaller than \code{row_km}, but this might means the
original \code{row_km} is not a good choice.
\item \code{row_split} Same as \code{split}.
\item \code{column_km} K-means clustering on columns.
\item \code{column_km_repeats} Number of k-means runs to get a consensus k-means clustering. Similar as \code{row_km_repeats}.
\item \code{column_split} Split on columns. For heatmap splitting, please refer to https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#heatmap-split .
\item \code{gap} Gap between row slices if the heatmap is split by rows. The value should be a \code{grid::unit} object.
\item \code{row_gap} Same as \code{gap}.
\item \code{column_gap} Gap between column slices.
\item \code{show_parent_dend_line} When heatmap is split, whether to add a dashed line to mark parent dendrogram and children dendrograms?
\item \code{width} Width of the heatmap body.
\item \code{height} Height of the heatmap body.
\item \code{heatmap_width} Width of the whole heatmap (including heatmap components)
\item \code{heatmap_height} Height of the whole heatmap (including heatmap components). Check https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#size-of-the-heatmap .
\item \code{show_heatmap_legend} Whether show heatmap legend?
\item \code{heatmap_legend_param} A list contains parameters for the heatmap legends. See \verb{color_mapping_legend,ColorMapping-method} for all available parameters.
\item \code{use_raster} Whether render the heatmap body as a raster image. It helps to reduce file size when the matrix is huge. If number of rows or columns is more than 2000, it is by default turned on. Note if \code{cell_fun}
is set, \code{use_raster} is enforced to be \code{FALSE}.
\item \code{raster_device} Graphic device which is used to generate the raster image.
\item \code{raster_quality} A value larger than 1.
\item \code{raster_device_param} A list of further parameters for the selected graphic
device. For raster image support, please check
https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#heatmap-as-raster-image.
\item \code{raster_resize_mat} Whether resize the matrix to let the dimension of the
matrix the same as the dimension of the raster image?  The value
can be logical. If it is \code{TRUE}, \code{base::mean} is used to
summarize the sub matrix which corresponds to a single pixel.  The
value can also be a summary function, e.g. \code{base::max}.
\item \code{raster_by_magick} Whether to use \code{magick::image_resize} to scale the
image.
\item \code{raster_magick_filter} Pass to \code{filter} argument of
\code{magick::image_resize}. A character scalar and all possible values are in
\code{magick::filter_types}. The default is \code{"Lanczos"}.
\item \code{post_fun} A function which will be executed after the heatmap list is
drawn.
}}

\item{ggparams}{Other arguments passed to \code{ggfn}.}

\item{debug}{In the event that it is set to TRUE, the
\link[ggplot2:ggplot]{ggplot} object gracefully yielded by ggfn, during the
execution of \link[ComplexHeatmap:draw-dispatch]{draw}, shall be returned directly.
Alternatively, one may opt to impart a function (whereas a formula would also
be deemed acceptable) to be carried out conjointly with the
\link[ggplot2:ggplot]{ggplot} object, as exemplified by \link{print} or \link{browser}.}
}
\value{
A \code{ggHeatmap} Object.
}
\description{
Plot heatmaps layer by layer
}
\details{
The initialization function only applies parameter checking and fill values to the slots with some validation.

Following methods can be applied to the \code{Heatmap-class} object:
\itemize{
\item \verb{show,ggHeatmap-method}: draw a single heatmap with default parameters
\item \verb{draw,ggHeatmap-method}: draw a single heatmap.
\item \code{+} or \verb{\%v\%} append heatmaps and annotations to a list of heatmaps.
}

The constructor function pretends to be a high-level graphic function because
the \code{show} method of the \code{Heatmap-class} object actually plots the
graphics.
}
\section{ggfn}{


\code{ggfn} accept a ggplot2 object with a default data and mapping created by
\code{ggplot(data, aes(.data$.column, .data$.row))}.
the data contains 6 columns:
\itemize{
\item \code{.slice_row}: the slice row number
\item \code{.slice_column}: the slice column number
\item \code{.row} and \code{.column}: the row and column coordinates
\item \code{.row_index} and \code{.column_index}: the row and column index of the original
matrix.
}
}

