% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno-subsettable-.R
\name{new_anno_subset}
\alias{new_anno_subset}
\title{Subsettable AnnotationFunction}
\usage{
new_anno_subset(
  matrix,
  draw_fn,
  ...,
  ylim = NULL,
  subset_rule = NULL,
  width = NULL,
  height = NULL,
  show_name = TRUE,
  which = NULL,
  name = NULL
)
}
\arguments{
\item{matrix}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{draw_fn}{A function which defines how to draw the annotation. The
function should accept at least \code{matrix}, \code{index}, \code{k}, \code{n} arguments.
\code{...} is also passed on to this function.}

\item{...}{Additional arguments passed on to \code{draw_fn}.}

\item{ylim}{Ranges of data value on the data axis.}

\item{subset_rule}{A list of function to subset variables in \code{...}.}

\item{width}{The width of the plotting region (the viewport) that the annotation is drawn. If it is a row annotation, the width must be an absolute unit. Since the \code{AnnotationFunction} object is always contained by the \code{\link[ComplexHeatmap]{SingleAnnotation-class}}object, you can only set the width of row annotations or height of column annotations, while e.g. the height of the row annotation is always \code{unit(1, "npc")} which means it always fully filled in the parent \code{SingleAnnotation} and only in \code{\link[ComplexHeatmap]{SingleAnnotation}} or even \code{\link[ComplexHeatmap]{HeatmapAnnotation}} can adjust the height of the row annotations.}

\item{height}{The height of the plotting region (the viewport) that the annotation is drawn. If it is a column annotation, the width must be an absolute unit.}

\item{show_name}{It is used to turn off the drawing of annotation names in \code{\link[ComplexHeatmap]{HeatmapAnnotation}}. Annotations always have names associated and normally they will be drawn beside the annotation graphics to tell what the annotation is about. e.g. the annotation names put beside the points annotation graphics. However, for some of the annotations, the names are not necessarily to be drawn, such as text annotations drawn by \code{\link[ComplexHeatmap]{anno_text}} or an empty annotation drawn by \code{\link[ComplexHeatmap]{anno_empty}}. In this case, when \code{show_names} is set to \code{FALSE}, there will be no annotation names drawn for the annotation.}

\item{which}{Whether it is drawn as a column annotation or a row annotation?}

\item{name}{Name of the annotation, only used for message.}
}
\value{
A \code{\link[ComplexHeatmap]{AnnotationFunction-class}} object which can be used in \code{\link[ComplexHeatmap]{HeatmapAnnotation}}.
}
\description{
Subsettable AnnotationFunction
}
\details{
\code{new_anno_subset} is similar with
\link[ComplexHeatmap:AnnotationFunction]{AnnotationFunction}, but
\code{new_anno_subset} won't change the function environment of \code{draw_fn}. So it's
safe to use \code{new_anno_subset} in pacakge development, particularly when
dealing with internal functions in the package namespace that are likely to
exist. You must always ensure arguments passed on to \code{...} have subset rules
since the internal will set \verb{@subsettable} to \code{TRUE}.
}
\examples{
anno <- new_anno_subset(
    rnorm(10L),
    draw_fn = function(matrix, index, k, n) {
        n <- length(index)
        pushViewport(viewport(xscale = c(0.5, n + 0.5), yscale = c(0, 10)))
        grid.rect()
        grid.points(1:n,
            matrix[index, , drop = FALSE],
            default.units = "native"
        )
        if (k == 1) grid.yaxis()
        popViewport()
    },
    height = unit(2, "cm")
)
draw(anno)
}
\seealso{
\itemize{
\item \link{new_anno}
\item \link[ComplexHeatmap:AnnotationFunction]{AnnotationFunction}
}
}
