% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R, R/prepare.R,
%   R/utils-complexheatmap.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{draw}
\alias{prepare}
\alias{ht_opt}
\alias{gpar}
\alias{unit}
\alias{alpha}
\alias{.pt}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ComplexHeatmap}{\code{\link[ComplexHeatmap:draw-dispatch]{draw}}, \code{\link[ComplexHeatmap]{ht_opt}}, \code{\link[ComplexHeatmap:prepare-Heatmap-method]{prepare}}}

  \item{ggplot2}{\code{\link[ggplot2:graphical-units]{.pt}}, \code{\link[ggplot2:reexports]{alpha}}}

  \item{grid}{\code{\link[grid]{gpar}}, \code{\link[grid]{unit}}}
}}

