% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R, R/utils-complexheatmap.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{draw}
\alias{ht_opt}
\alias{gpar}
\alias{unit}
\title{Objects exported from other packages}
\examples{
draw(ggheat(matrix(rnorm(81), nrow = 9)))
ht_opt("heatmap_row_names_gp")
gpar(col = "red")
unit(1, "npc")
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ComplexHeatmap}{\code{\link[ComplexHeatmap:draw-dispatch]{draw}}, \code{\link[ComplexHeatmap]{ht_opt}}}

  \item{grid}{\code{\link[grid]{gpar}}, \code{\link[grid]{unit}}}
}}

