% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gpar}
\alias{unit}
\title{Objects exported from other packages}
\value{
\itemize{
\item \code{gpar}: An object of class \link[grid:gpar]{gpar}.
}

\itemize{
\item \code{unit}: An object of class \link[grid:unit]{unit}.
}
}
\examples{
gpar(col = "red")
unit(1, "npc")
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grid}{\code{\link[grid]{gpar}}, \code{\link[grid]{unit}}}
}}

